atests/requirements.txt
robotframework==3.1.2
robotframework-requests==0.6.5
psycopg2-binary==2.8.4